requirements.txt
arrow==1.4.0
asgiref==3.11.0
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
Django==6.0
django-browser-reload==1.21.0
django-tailwind==4.4.2
djangorestframework==3.16.1
easyocr==1.7.2
filelock==3.20.0
fsspec==2025.12.0
honcho==2.0.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
pyclipper==1.4.0
Pygments==2.19.2
pytailwindcss==0.3.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sqlparse==0.5.4
sympy==1.14.0
text-unidecode==1.3
tifffile==2025.10.16
torch==2.9.1
torchvision==0.24.1
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0